Polished code

+++ RRCCommon/EigenVectorXdUtil.cpp
@@ -4,7 +4,7 @@ EigenVectorXdUtil::EigenVectorXdUtil() {
 
 }
 
-float EigenVectorXdUtil::max(const Eigen::VectorXd &vec) {
+float EigenVectorXdUtil::max(const Eigen::VectorXf &vec) {
     if (vec.size() == 0) throw std::invalid_argument("Calling to max of VectorXd, vector can not be empty");
 
     float maxElement = vec(0);
